shop/serializers.py
from rest_framework.serializers import ModelSerializer
from rest_framework import serializers

from .models import Application
from phonenumber_field.serializerfields import PhoneNumberField

from django.db import transaction