requirements.txt
django==4.2.1
environs==9.5.0
django-phonenumber-field[phonenumberslite]
dj-database-url==2.0.0
Pillow==9.5.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0
